SS: Add sun8iw11p1 support.
1. Add AES-CFB/OFB mode support;
2. Add SHA384/512 support;
3. Fix the compile waring;
4. Fix the clk config in dts.

+++ drivers/crypto/sunxi/sunxi_ss_proc_comm.c
@@ -162,8 +162,8 @@ int ss_hash_padding(ss_hash_ctx_t *ctx, int type)
 		*(int *)(p+4) = len_h;
 	}
 
-	SS_DBG("After padding %ld: %02x %02x %02x %02x   %02x %02x %02x %02x\n",
-			p + 8 - ctx->pad,
+	SS_DBG("After padding %d: %02x %02x %02x %02x   %02x %02x %02x %02x\n",
+			(s32)(p + 8 - ctx->pad),
 			p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7]);
 
 	return p + 8 - ctx->pad;
